Github profile component, new dependency and minor changes.
Added the github profile component. If a github username is specified
in the component, a widget is drawn. It looks pretty nice. The page is
coming together beautifully. Angular is amazing!!!
Removed the mygithub from the navbar as that section will no longer
be implemented. Also removed the component files related to it.